hashmap implementation and runner
