+++ src/com/akash/basic/Runner3.java
@@ -1,32 +1,26 @@
 package com.akash.basic;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import com.akash.basic.Runner1.MockApp;
-
-public class Runner3{
-
-    static class MyKey {
-        int value;
-        MyKey(int value) { this.value = value; }
-    
-        public boolean equals(Object o) {
-            return true;
-        }
-    
-        public int hashCode() {
-            return 0;
-        }
-    }
+public class Runner3 {
 
     public static void main(String[] args) {
-        Map<MyKey, String> map = new HashMap<>();
-        map.put(new MyKey(1), "One");
-        map.put(new MyKey(2), "Two");
-        System.out.println(map.size()); // 1
+        
+        new Y().method(100);
     }
-    
+
+    static class X{
+        void method(int a){
+            System.out.println("ONE");
+        }
+     
+        void method(double d){
+            System.out.println("TWO");
+        }
+    }
+     
+    static class Y extends X{
+        @Override
+        void method(double d){
+            System.out.println("THREE");
+        }
+    }
 }
-
-
