+++ src/com/akash/basic/hashMap/Mock6
