+++ src/com/akash/basic/hashMap/Runner8.java
@@ -1,4 +1,4 @@
-package com.akash.basic;
+package com.akash.basic.hashMap;
 
 import java.util.HashMap;
 import java.util.Map;
