+++ src/com/akash/basic/hashMap/mock3
@@ -1,0 +1,21 @@
+class MyKey {
+    int value;
+    MyKey(int value) { this.value = value; }
+
+    public boolean equals(Object o) {
+        return true;
+    }
+
+    public int hashCode() {
+        return 0;
+    }
+}
+
+public class Main {
+    public static void main(String[] args) {
+        Map<MyKey, String> map = new HashMap<>();
+        map.put(new MyKey(1), "One");
+        map.put(new MyKey(2), "Two");
+        System.out.println(map.size()); // 2
+    }
+}
