--- src/com/akash/basic/mock12
@@ -1,25 +1,0 @@
-class X{
-    void method(int a){
-        System.out.println("ONE");
-    }
- 
-    void method(double d){
-        System.out.println("TWO");
-    }
-}
- 
-class Y extends X{
-    @Override
-    void method(double d)
-    {
-        System.out.println("THREE");
-    }
-}
-
-
-public class Main {
-  public static void main(String[] args)
-    {
-        new Y().method(100);
-    }
-}
