+++ src/com/akash/basic/mock2
@@ -1,23 +1,31 @@
-class Employee {
-    String name;
-    Employee(String name) {
-        this.name = name;
+public class OverloadingExample {
+    public void print(int value) {
+        System.out.println("Primitive int method called with value: " + value);
     }
 
-    public int hashCode() {
-        return 42;
+    public void print(Integer value) {
+        System.out.println("Integer object method called with value: " + value);
     }
-}
 
-public class Main {
+    public void print(object value){
+        System.out.println("object method called with value: " + value);
+    }
+
     public static void main(String[] args) {
-        Map<Employee, String> map = new HashMap<>();
-        Employee e1 = new Employee("Amit");
-        Employee e2 = new Employee("Amit");
+        OverloadingExample example = new OverloadingExample();
+        
+        // Calling with primitive int
+        example.print(10); // Output: Primitive int method called with value: 10
+        
+        // Calling with Integer object
+        example.print(Integer.valueOf(20)); // Output: Integer object method called with value: 20
+        
+        // Calling with autoboxed int
+        example.print(30); // Output: Primitive int method called with value: 30
+        
+        // Calling with autoboxed Integer
+        example.print(new Integer(40)); // Output: Integer object method called with value: 40
 
-        map.put(e1, "Developer");
-        map.put(e2, "Manager"); // store 
-
-        System.out.println(map.get(e1)); // manager
+        example.print(210);
     }
 }
