+++ src/com/akash/basic/stream/mock1
@@ -1,0 +1,1 @@
+print 1 to 100 number using stream
